Remove messages instead of emptying the container.
There may be additional content in the flash-messages container,
especially if the user is using Bootstrap and does not define a
.unobtrusive-flash-container element. Therefore, instead of emptying the
messages container, delete the message elements by matching their CSS
class.
NOTE: With Bootstrap, all .alert elements in the messages container will
be cleared before a new message is shown when
UnobtrusiveFlash.flashOptions.clearFlashOnNextRequest is truthy.