Better handle `cacheable-flash` transitioners
The `cacheable-flash` gem is no longer maintained and the author suggests `unobtrusive_flash` as its successor: https://github.com/pboling/cacheable-flash

However, the flash cookie format is different between these two libraries. This patch will make the transition from one to the other more painless (and as an added side effect also to help to handle tampered cookies that may not be in the expected format). I manually did this locally for our project and figured others might benefit from it as well.