Slight code re-organization
The current organization of this gem makes it difficult to customize the behavior of this gem because it contains explicit references to fully qualified method: `UnobtrusiveFlash::ControllerMixin.append_flash_to_cookie` (which we had a need to tweak slightly).

I reorganized to make this unneeded and also to follow a common module pattern in the ruby community. This change was beneficial to my team and I suspect might be appreciated by others. It allowed us to easily override methods as needed and the code remains easily testable.